pyproject.toml
[project]
name = "ai-voice-coach-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datasets>=4.3.0",
    "faster-whisper>=1.2.1",
    "ipython>=9.6.0",
    "librosa>=0.11.0",
    "matplotlib>=3.10.7",
    "openai-whisper>=20250625",
    "pandas>=2.3.3",
    "pip>=25.3",
    "soundfile>=0.13.1",
    "torch<2.9.0",
    "torchaudio<2.9.0",
    "torchcodec==0.7",
    "tqdm>=4.67.1",
    "transformers>=4.57.1",
]
